# new kernel for graphSum