Code: https://github.com/Liu-rj/gs-experiments/tree/master/dgl_e2e_benchmark/ladies

Device: 1080Ti

| Dataset/Time Cost per epoch [sec] | Before  | After   |
| --------------------------------- | ------- | ------- |
| reddit                            | 13.1663 | 11.7882 |
| products                          | 27.9617 | 11.4949 |