@Liu-rj 2022-11-29
It seems that there's still a gap of sampling time between ours and dal's version?