@gpzlx1 2022-11-29
> It seems that there's still a gap of sampling time between ours and dgl's version?
> 
> ---------------Edited Oh, I forget that you test it on 1080Ti, would you also check dgl's sampling time?( I'm concerned about whether we need to optimize our implementation of `COOSort`.)

In my computer, Ours is faster than DGL.