@gpzlx1 2022-11-29
> A unit test is needed.

GraphSum has had test files before.
In [test_graph_compute.cc](https://github.com/czkkkkkk/graph_sampling/blob/master/tests/test_graph_compute.cc).